requirements.txt
aircv==1.4.6
certifi==2021.5.30
cycler==0.10.0
keyboard==0.13.5
kiwisolver==1.3.2
matplotlib==3.4.3
MouseInfo==0.1.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
six==1.16.0
wincertstore==0.2
